LPD-97871 Extract the shared FDS asset list props into their own module
The Recent Assets props transformer carried the whole behavior of a CMS asset list: the item actions with their visibility filters, the action click handler with every modal it opens, and the cell renderer. None of that is specific to Recent Assets, so it now lives in its own module and the transformer is the thin adapter the taglib needs.
